스프링 시큐리티 포스트

+++ src/main/java/com/sparta/magazine_week2/controller/PostController.java
@@ -1,9 +1,6 @@
 package com.sparta.magazine_week2.controller;
 
-import com.sparta.magazine_week2.dto.PostRemoveRequestDto;
-import com.sparta.magazine_week2.dto.PostRequestDto;
-import com.sparta.magazine_week2.dto.PostUpdateRequestDto;
-import com.sparta.magazine_week2.dto.UserResponseDto;
+import com.sparta.magazine_week2.dto.*;
 import com.sparta.magazine_week2.entity.Post;
 import com.sparta.magazine_week2.repository.PostRepository;
 import com.sparta.magazine_week2.security.UserDetailsImpl;
@@ -24,9 +21,15 @@ public class PostController {
     }
 
     @PostMapping("/api/post") //게시물 등록
-    public UserResponseDto createpost(@RequestBody PostRequestDto requestDto) {
+    public UserResponseDto createpost(@RequestBody PostRequestDto requestDto, @AuthenticationPrincipal UserDetailsImpl userDetails) {
         UserResponseDto responseDto = new UserResponseDto();
         Post post = new Post(requestDto);
+        String username = userDetails.getUsername();
+        if (username == null) {
+            responseDto.setResult(false);
+            responseDto.setMsg("로그인 후 등록이 가능합니다.");
+            return responseDto;
+        }
         postRepository.save(post);
         responseDto.setResult(true);
         responseDto.setMsg("success");
@@ -36,8 +39,13 @@ public class PostController {
 
     //좋아요 다만들면 리턴 좋아요도 같이 해주기
     @GetMapping("/api/post") //게시물 조회
-    public List<Post> getPost(){
-        return postRepository.findAllByOrderByModifiedAtDesc();
+    public PostResponseDto getPost(@AuthenticationPrincipal UserDetailsImpl userDetails) {
+        PostResponseDto responseDto = new PostResponseDto();
+        String username = userDetails.getUsername();
+        List<Post> posts = postRepository.findAllByOrderByModifiedAtDesc();
+        responseDto.setPosts(posts);
+        responseDto.setUsername(username);
+        return responseDto;
     }
 
     //게시물 수정
@@ -49,14 +57,22 @@ public class PostController {
         if (!username.equals(username2)) {
             responseDto.setResult(false);
             responseDto.setMsg("작성자만 수정이 가능합니다.");
+            return responseDto;
         }
         return postService.update(updateRequestDto);
     }
 
     //게시물 삭제
     @DeleteMapping("/api/post") // 게시물 삭제
-    public UserResponseDto deletePost(@RequestBody PostRemoveRequestDto removeRequestDto, @AuthenticationPrincipal UserDetailsImpl userDtails){
+    public UserResponseDto deletePost(@RequestBody PostRemoveRequestDto removeRequestDto, @AuthenticationPrincipal UserDetailsImpl userDetails){
         UserResponseDto responseDto = new UserResponseDto();
+        String username = userDetails.getUsername();
+        String username2 = removeRequestDto.getUsername();
+        if (!username.equals(username2)) {
+            responseDto.setResult(false);
+            responseDto.setMsg("작성자만 삭제가 가능합니다.");
+            return responseDto;
+        }
         postRepository.deleteById(removeRequestDto.getPostId());
         responseDto.setResult(true);
         responseDto.setMsg("success");
